refactor: Remove custom GOAP and integrate dogoap library
- Deleted all custom GOAP implementation files (goap_actions.rs, goap_planner.rs, goap_bridge.rs, goap_state_sync.rs, goap_loader.rs)
- Removed all references to ActionPlan, GoapAction, StateValue, WorldState, ActionSet types
- Integrated bevy_dogoap library with basic actions (EatAction, RestAction, GatherFoodAction)
- Created bevy_dogoap_impl.rs with dogoap-based AI implementation
- Updated all systems to remove dependencies on custom GOAP types
- Fixed compilation errors by cleaning up imports and unused functions
- Project now builds successfully with dogoap integration

Note: Runtime configuration of dogoap planner still needs refinement, but custom GOAP has been fully removed.

+++ world_sim_simple/src/ai/bevy_dogoap_impl.rs
@@ -19,13 +19,13 @@ pub struct RestAction;
 pub struct GatherFoodAction;
 
 // State components using DatumComponent
-#[derive(Component, Clone)]
-pub struct Hunger(pub f32);
+#[derive(Component, Clone, DatumComponent)]
+pub struct Hunger(pub f64);
 
-#[derive(Component, Clone)]
-pub struct Energy(pub f32);
+#[derive(Component, Clone, DatumComponent)]
+pub struct Energy(pub f64);
 
-#[derive(Component, Clone)]
+#[derive(Component, Clone, DatumComponent)]
 pub struct FoodCount(pub f64);
 
 // System to handle EatAction execution
@@ -121,7 +121,7 @@ pub fn setup_dogoap_planners(
         
         // Define actions with their preconditions and effects
         let eat_action = EatAction::new()
-            .add_precondition(FoodCount::is_more(0))
+            .add_precondition(FoodCount::is_more(0.0))
             .add_mutator(Hunger::increase(10.0))  // Matches what handle_eat_action actually does
             .add_mutator(FoodCount::decrease(1.0));
